moved aggregation actions into pkg/common/aggregation
