+++ pkg/common/aggregation/aggregateAzurePeering.go
@@ -1,4 +1,4 @@
-package reconcile
+package aggregation
 
 import (
 	"context"
